# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API Documentation",
        "description": "API for aggregating news from multiple sources with personalized feeds",
        "contact": {
            "email": "sherifanwar766@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://news-aggregator.test",
            "description": "Local Development Server"
        },
        {
            "url": "http://news-aggregator.com",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/v1/user/feed": {
            "get": {
                "tags": [
                    "User Feed"
                ],
                "summary": "Get personalized article feed",
                "description": "Retrieve articles based on user preferences (sources, categories, authors). Query parameters can override or extend preferences.",
                "operationId": "233d2def980dc1f128aea47dba5b7271",
                "parameters": [
                    {
                        "name": "searchTerm",
                        "in": "query",
                        "description": "Search term (extends user preferences)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "technology"
                        }
                    },
                    {
                        "name": "source[]",
                        "in": "query",
                        "description": "Additional sources to include (merges with preferred sources)",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "bbc"
                            }
                        }
                    },
                    {
                        "name": "category[]",
                        "in": "query",
                        "description": "Additional categories to include (merges with preferred categories)",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "science"
                            }
                        }
                    },
                    {
                        "name": "from_date",
                        "in": "query",
                        "description": "Filter articles from this date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-10-01"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "description": "Filter articles until this date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-10-21"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Override user's default sort field",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "published_at",
                                "created_at",
                                "title"
                            ],
                            "example": "published_at"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Override user's default sort order",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "example": "desc"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Override user's default items per page (1-100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "example": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Personalized feed retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Latest Technology News"
                                                    },
                                                    "source": {
                                                        "type": "object"
                                                    },
                                                    "category": {
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "personalization": {
                                            "properties": {
                                                "personalized": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "filters_applied": {
                                                    "properties": {
                                                        "sources": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string",
                                                                "example": "guardian"
                                                            }
                                                        },
                                                        "categories": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string",
                                                                "example": "technology"
                                                            }
                                                        },
                                                        "has_author_filter": {
                                                            "type": "boolean",
                                                            "example": true
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "user_preferences": {
                                                    "properties": {
                                                        "articles_per_page": {
                                                            "type": "integer",
                                                            "example": 20
                                                        },
                                                        "default_sort": {
                                                            "type": "string",
                                                            "example": "published_at"
                                                        },
                                                        "default_order": {
                                                            "type": "string",
                                                            "example": "desc"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}